// src/modules/apartment/enums/heating-type.enum.ts
export enum HeatingType {
  Central = 'central', // etazno
  District = 'district', // centralno
  Electricity = 'electricity',
  Gas = 'gas',
  MarbleRadiators = 'marble radiators',
  NorwegianRadiators = 'norwegian radiators',
  StorageHeater = 'storage heater', // TA pec
  ThermalPump = 'thermal pump',
  TileStove = 'tile stove',
  Underfloor = 'underfloor',
}